Improvements of User profile

+++ Carpool/Carpool.App/ViewModel/ProfileUserDetailViewModel.cs
@@ -1,20 +1,25 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Input;
 using Carpool.App.Command;
 using Carpool.App.Messages;
 using Carpool.App.Services;
 using Carpool.App.Wrapper;
 using Carpool.BL.Facades;
+using Carpool.BL.Models;
+using Microsoft.Win32;
 
 namespace Carpool.App.ViewModel
 {
     public class ProfileUserDetailViewModel : UserDetailViewModelBase, IProfileUserDetailViewModel
     {
         private readonly ISession _session;
+
 
         public ProfileUserDetailViewModel(
             UserFacade userFacade, 
@@ -28,7 +33,7 @@ namespace Carpool.App.ViewModel
             SaveChangesCommand = new AsyncRelayCommand(OnSaveChanges, CanSave);
             DeleteAccountCommand = new AsyncRelayCommand(OnDeleteAccount);
             LogOutCommand = new RelayCommand(OnLogOut);
-            DisplayCarEditCommand = new AsyncRelayCommand(OnDisplayCarEdit);
+            DisplayCarEditCommand = new RelayCommand(OnDisplayCarEdit);
             DisplayRideListCommand = new RelayCommand(OnDisplayRideList);
         }
 
@@ -49,7 +54,6 @@ namespace Carpool.App.ViewModel
             return Model is {HasErrors: false};
         }
 
-
         private async Task OnSaveChanges()
         {
             await SaveAsync();
@@ -58,22 +62,32 @@ namespace Carpool.App.ViewModel
 
         private async Task OnDeleteAccount()
         {
-            await DeleteAsync();
-            Mediator.Send(new DisplayLoginScreenMessage());
+            var answer = MessageBox.Show(
+                "The " + Model?.Name + " " + Model?.Surname + " will be deleted forever.\nAre you sure?",
+                "Delete account",
+                MessageBoxButton.YesNo,
+                MessageBoxImage.Warning,
+                MessageBoxResult.No);
+
+            if (answer == MessageBoxResult.Yes && Model != null)
+            {
+                await DeleteAsync();
+                Mediator.Send(new DisplayLoginScreenMessage());
+            }
         }
 
 
         private void OnLogOut()
         {
-            System.Diagnostics.Debug.WriteLine("Log Out...");
             _session.LogUserOut();
             Mediator.Send(new DisplayLoginScreenMessage());
+            Model = GetEmptyUser();
         }
 
 
-        private async Task OnDisplayCarEdit()
+        private void OnDisplayCarEdit()
         {
-            await LoadDefaultProfile();
+
         }
 
 
@@ -81,7 +95,6 @@ namespace Carpool.App.ViewModel
         {
 
         }
-
 
         private async void OnLoadUserProfileMessage(LoadUserProfileMessage message)
         {
@@ -93,6 +106,8 @@ namespace Carpool.App.ViewModel
             {
                 await LoadDefaultProfile();
             }
+
+            OnPropertyChanged();
         }
 
 
@@ -109,8 +124,6 @@ namespace Carpool.App.ViewModel
             {
                 Model = GetEmptyUser();
             }
-
-            OnPropertyChanged();
         }
     }
 }
